pyproject.toml
[tool.poetry]
name = "wb-parse"
version = "0.1.0"
description = ""
authors = ["Lapitski Alexei <alexeylapitski@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openpyxl = "^3.1.2"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
selenium = "^4.18.1"
aiogram = "^3.4.1"
pydantic = "<=2.5"
celery = "^5.3.6"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
